Re-namespace statement component resolvers
