+++ src/StatementComponentResolver/StatementComponentResolverInterface.php
@@ -2,11 +2,14 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilResolver;
+namespace webignition\BasilResolver\StatementComponentResolver;
 
 use webignition\BasilModelProvider\Exception\UnknownItemException;
 use webignition\BasilModelProvider\ProviderInterface;
 use webignition\BasilModels\StatementInterface;
+use webignition\BasilResolver\ResolvedComponentInterface;
+use webignition\BasilResolver\UnknownElementException;
+use webignition\BasilResolver\UnknownPageElementException;
 
 interface StatementComponentResolverInterface
 {
