+++ src/StatementComponentResolver/StatementValueUrlResolver.php
@@ -2,14 +2,17 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilResolver;
+namespace webignition\BasilResolver\StatementComponentResolver;
 
 use webignition\BasilModelProvider\Exception\UnknownItemException;
 use webignition\BasilModelProvider\ProviderInterface;
 use webignition\BasilModels\PageProperty\PageProperty;
 use webignition\BasilModels\StatementInterface;
+use webignition\BasilResolver\ImportedUrlResolver;
+use webignition\BasilResolver\ResolvedComponent;
+use webignition\BasilResolver\ResolvedComponentInterface;
 
-class StatementIdentifierUrlResolver implements StatementComponentResolverInterface
+class StatementValueUrlResolver implements StatementComponentResolverInterface
 {
     public function __construct(
         private ImportedUrlResolver $importedUrlResolver
@@ -18,7 +21,7 @@ class StatementIdentifierUrlResolver implements StatementComponentResolverInterf
 
     public static function createResolver(): self
     {
-        return new StatementIdentifierUrlResolver(
+        return new StatementValueUrlResolver(
             ImportedUrlResolver::createResolver()
         );
     }
@@ -31,19 +34,18 @@ class StatementIdentifierUrlResolver implements StatementComponentResolverInterf
         ProviderInterface $pageProvider,
         ProviderInterface $identifierProvider
     ): ?ResolvedComponentInterface {
-        $identifier = $statement->getIdentifier();
+        $value = $statement->getValue();
+        if (is_string($value) && false === PageProperty::is($value)) {
+            $resolvedValue = $this->importedUrlResolver->resolve($value, $pageProvider);
 
-        if (is_string($identifier) && false === PageProperty::is($identifier)) {
-            $resolvedIdentifier = $this->importedUrlResolver->resolve($identifier, $pageProvider);
-
-            if ($identifier !== $resolvedIdentifier) {
-                $resolvedIdentifier = '"' . $resolvedIdentifier . '"';
+            if ($value !== $resolvedValue) {
+                $resolvedValue = '"' . $resolvedValue . '"';
             }
 
             return new ResolvedComponent(
-                ResolvedComponentInterface::TYPE_IDENTIFIER,
-                $identifier,
-                $resolvedIdentifier
+                ResolvedComponentInterface::TYPE_VALUE,
+                $value,
+                $resolvedValue
             );
         }
